Add logger from common-src

See merge request oai/cn5g/oai-cn5g-smf!186

+++ src/smf_app/smf_procedure.cpp
@@ -1253,7 +1253,7 @@ smf_procedure_code session_update_sm_context_procedure::run(
     default: {
       Logger::smf_app().error(
           "Update SM Context procedure: Unknown session management type %d",
-          session_procedure_type);
+          (int) session_procedure_type);
     }
   }
